Sparky2: allow setting the LPF for accel and gyro
This code was previously disabled. Added an option for each
of them and reenabled the code.

Note, it does not appear possible to use a combination of
settings that allows downsampling the data from 8khz, regardless
of LPF setting. Putting the LPF to 250 Hz will result in the
data rate being 8khz, regardless of the sampling setting.